move functions to abstract class
