+++ sup3r/models/base.py
@@ -10,8 +10,6 @@ import pandas as pd
 import tensorflow as tf
 from tensorflow.keras import optimizers
 from warnings import warn
-from rex.utilities.utilities import safe_json_load
-from phygnn import CustomNetwork
 
 from sup3r.models.abstract import AbstractInterface, AbstractSingleModel
 from sup3r.utilities import VERSION_RECORD
@@ -158,235 +156,6 @@ class Sup3rGan(AbstractInterface, AbstractSingleModel):
         params = cls.load_saved_params(model_dir, verbose=verbose)
 
         return cls(fp_gen, fp_disc, **params)
-
-    def load_network(self, model, name):
-        """Load a CustomNetwork object from hidden layers config, .json file
-        config, or .pkl file saved pre-trained model.
-
-        Parameters
-        ----------
-        model : str | dict
-            Model hidden layers config, a .json with "hidden_layers" key, or a
-            .pkl for a saved pre-trained model.
-        name : str
-            Name of the model to be loaded
-
-        Returns
-        -------
-        model : phygnn.CustomNetwork
-            CustomNetwork object initialized from the model input.
-        """
-
-        if isinstance(model, str) and model.endswith('.json'):
-            model = safe_json_load(model)
-            self._meta[f'config_{name}'] = model
-            if 'hidden_layers' in model:
-                model = model['hidden_layers']
-            elif ('meta' in model
-                  and f'config_{name}' in model['meta']
-                  and 'hidden_layers' in model['meta'][f'config_{name}']):
-                model = model['meta'][f'config_{name}']['hidden_layers']
-            else:
-                msg = ('Could not load model from json config, need '
-                       '"hidden_layers" key or '
-                       f'"meta/config_{name}/hidden_layers" '
-                       ' at top level but only found: {}'
-                       .format(model.keys()))
-                logger.error(msg)
-                raise KeyError(msg)
-
-        elif isinstance(model, str) and model.endswith('.pkl'):
-            model = CustomNetwork.load(model)
-
-        if isinstance(model, list):
-            model = CustomNetwork(hidden_layers=model, name=name)
-
-        if not isinstance(model, CustomNetwork):
-            msg = ('Something went wrong. Tried to load a custom network '
-                   'but ended up with a model of type "{}"'
-                   .format(type(model)))
-            logger.error(msg)
-            raise TypeError(msg)
-
-        return model
-
-    @property
-    def means(self):
-        """Get the data normalization mean values.
-
-        Returns
-        -------
-        np.ndarray
-        """
-        return self._means
-
-    @property
-    def stdevs(self):
-        """Get the data normalization standard deviation values.
-
-        Returns
-        -------
-        np.ndarray
-        """
-        return self._stdevs
-
-    def set_norm_stats(self, new_means, new_stdevs):
-        """Set the normalization statistics associated with a data batch
-        handler to model attributes.
-
-        Parameters
-        ----------
-        new_means : list | tuple | np.ndarray
-            1D iterable of mean values with same length as number of features.
-        new_stdevs : list | tuple | np.ndarray
-            1D iterable of stdev values with same length as number of features.
-        """
-
-        if self._means is not None:
-            logger.info('Setting new normalization statistics...')
-            logger.info("Model's previous data mean values: {}"
-                        .format(self._means))
-            logger.info("Model's previous data stdev values: {}"
-                        .format(self._stdevs))
-
-        self._means = new_means
-        self._stdevs = new_stdevs
-
-        if not isinstance(self._means, np.ndarray):
-            self._means = np.array(self._means)
-        if not isinstance(self._stdevs, np.ndarray):
-            self._stdevs = np.array(self._stdevs)
-
-        logger.info('Set data normalization mean values: {}'
-                    .format(self._means))
-        logger.info('Set data normalization stdev values: {}'
-                    .format(self._stdevs))
-
-    def norm_input(self, low_res):
-        """Normalize low resolution data being input to the generator.
-
-        Parameters
-        ----------
-        low_res : np.ndarray
-            Un-normalized low-resolution input data in physical units, usually
-            a 4D or 5D array of shape:
-            (n_obs, spatial_1, spatial_2, n_features)
-            (n_obs, spatial_1, spatial_2, n_temporal, n_features)
-
-        Returns
-        -------
-        low_res : np.ndarray
-            Normalized low-resolution input data, usually a 4D or 5D array of
-            shape:
-            (n_obs, spatial_1, spatial_2, n_features)
-            (n_obs, spatial_1, spatial_2, n_temporal, n_features)
-        """
-        if self._means is not None:
-            if isinstance(low_res, tf.Tensor):
-                low_res = low_res.numpy()
-
-            low_res = low_res.copy()
-            for idf in range(low_res.shape[-1]):
-                low_res[..., idf] -= self._means[idf]
-
-                if self._stdevs[idf] != 0:
-                    low_res[..., idf] /= self._stdevs[idf]
-                else:
-                    msg = ('Standard deviation is zero for '
-                           f'{self.training_features[idf]}')
-                    logger.warning(msg)
-                    warn(msg)
-
-        return low_res
-
-    def un_norm_output(self, hi_res):
-        """Un-normalize synthetically generated output data to physical units
-
-        Parameters
-        ----------
-        hi_res : tf.Tensor | np.ndarray
-            Synthetically generated high-resolution data
-
-        Returns
-        -------
-        hi_res : np.ndarray
-            Synthetically generated high-resolution data
-        """
-        if self._means is not None:
-            if isinstance(hi_res, tf.Tensor):
-                hi_res = hi_res.numpy()
-
-            for idf in range(hi_res.shape[-1]):
-                feature_name = self.output_features[idf]
-                i = self.training_features.index(feature_name)
-                mean = self._means[i]
-                stdev = self._stdevs[i]
-                hi_res[..., idf] = (hi_res[..., idf] * stdev) + mean
-
-        return hi_res
-
-    @property
-    def optimizer(self):
-        """Get the tensorflow optimizer to perform gradient descent
-        calculations for the generative network. This is functionally identical
-        to optimizer_disc is no special optimizer model or learning rate was
-        specified for the disc.
-
-        Returns
-        -------
-        tf.keras.optimizers.Optimizer
-        """
-        return self._optimizer
-
-    @property
-    def history(self):
-        """
-        Model training history DataFrame (None if not yet trained)
-
-        Returns
-        -------
-        pandas.DataFrame | None
-        """
-        return self._history
-
-    @property
-    def generator(self):
-        """Get the generative model.
-
-        Returns
-        -------
-        phygnn.base.CustomNetwork
-        """
-        return self._gen
-
-    @property
-    def generator_weights(self):
-        """Get a list of layer weights and bias terms for the generator model.
-
-        Returns
-        -------
-        list
-        """
-        return self.generator.weights
-
-    def _needs_lr_exo(self, low_res):
-        """Determine whether or not the sup3r model needs low-res exogenous
-        data
-
-        Parameters
-        ----------
-        low_res : np.ndarray
-            Low-resolution input data, usually a 4D or 5D array of shape:
-            (n_obs, spatial_1, spatial_2, n_features)
-            (n_obs, spatial_1, spatial_2, n_temporal, n_features)
-
-        Returns
-        -------
-        needs_lr_exo : bool
-            True if the model requires low-resolution exogenous data.
-        """
-
-        return low_res.shape[-1] < len(self.training_features)
 
     def generate(self, low_res, norm_in=True, un_norm_out=True,
                  exogenous_data=None):
@@ -1173,79 +942,3 @@ class Sup3rGan(AbstractInterface, AbstractSingleModel):
 
             if stop:
                 break
-
-    def finish_epoch(self, epoch, epochs, t0, loss_details,
-                     checkpoint_int, out_dir,
-                     early_stop_on, early_stop_threshold,
-                     early_stop_n_epoch, extras=None):
-        """Perform finishing checks after an epoch is done training
-
-        Parameters
-        ----------
-        epoch : int
-            Epoch number that is finishing
-        epochs : list
-            List of epochs being iterated through
-        t0 : float
-            Starting time of training.
-        loss_details : dict
-            Namespace of the breakdown of loss components
-        checkpoint_int : int | None
-            Epoch interval at which to save checkpoint models.
-        out_dir : str
-            Directory to save checkpoint models. Should have {epoch} in
-            the directory name. This directory will be created if it does not
-            already exist.
-        early_stop_on : str | None
-            If not None, this should be a column in the training history to
-            evaluate for early stopping (e.g. validation_loss_gen,
-            validation_loss_disc). If this value in this history decreases by
-            an absolute fractional relative difference of less than 0.01 for
-            more than 5 epochs in a row, the training will stop early.
-        early_stop_threshold : float
-            The absolute relative fractional difference in validation loss
-            between subsequent epochs below which an early termination is
-            warranted. E.g. if val losses were 0.1 and 0.0998 the relative
-            diff would be calculated as 0.0002 / 0.1 = 0.002 which would be
-            less than the default thresold of 0.01 and would satisfy the
-            condition for early termination.
-        early_stop_n_epoch : int
-            The number of consecutive epochs that satisfy the threshold that
-            warrants an early stop.
-        extras : dict | None
-            Extra kwargs/parameters to save in the epoch history.
-
-        Returns
-        -------
-        stop : bool
-            Flag to early stop training.
-        """
-
-        self.log_loss_details(loss_details)
-
-        self._history.at[epoch, 'elapsed_time'] = time.time() - t0
-        for key, value in loss_details.items():
-            if key != 'n_obs':
-                self._history.at[epoch, key] = value
-
-        last_epoch = epoch == epochs[-1]
-        chp = checkpoint_int is not None and (epoch % checkpoint_int) == 0
-        if last_epoch or chp:
-            msg = ('Model output dir for checkpoint models should have '
-                   f'{"{epoch}"} but did not: {out_dir}')
-            assert '{epoch}' in out_dir, msg
-            self.save(out_dir.format(epoch=epoch))
-
-        stop = False
-        if early_stop_on is not None and early_stop_on in self._history:
-            stop = self.early_stop(self._history, early_stop_on,
-                                   threshold=early_stop_threshold,
-                                   n_epoch=early_stop_n_epoch)
-            if stop:
-                self.save(out_dir.format(epoch=epoch))
-
-        if extras is not None:
-            for k, v in extras.items():
-                self._history.at[epoch, k] = v
-
-        return stop
